BugFix: using too many values on non-zero levels

+++ src/multitau.h
@@ -9,12 +9,12 @@ class MultiTauAutocorrelator {
         unsigned ntime_;
         unsigned len_;
         unsigned idx_;
-        float * signal_;
-        float * g2s_;
-        float * tau_; 
+        double * signal_;
+        double * g2s_;
+        double * tau_; 
         
     public:
-        MultiTauAutocorrelator(float *, unsigned, unsigned);
+        MultiTauAutocorrelator(double *, unsigned, unsigned);
         ~MultiTauAutocorrelator();
 
         /* forbid copy and assignment operators */
@@ -23,13 +23,13 @@ class MultiTauAutocorrelator {
 
         // expose g2 and lag-time arrays
         size_t length() const { return len_; }
-        float * g2()  { return g2s_; }
-        float * tau() { return tau_; }
+        double * g2()  { return g2s_; }
+        double * tau() { return tau_; }
 
         /* work horses */
         void reduceInplace();
         void run_level(unsigned);
-        void process(unsigned m=16, float dt=1);
+        void process(unsigned m=16, double dt=1);
 };
 
 #endif // MULTI_TAU_CORRELATOR__H
